LF-4765 Add addonPartner service
